requirements.txt
annotated-types==0.7.0
anyio==4.10.0
asgiref==3.9.1
beautifulsoup4==4.13.5
cachetools==5.5.2
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
contextlib2==21.6.0
cryptography==45.0.7
Django==5.2.5
django-ninja==1.4.3
django-ninja-extra==0.30.1
django-ninja-jwt==5.3.7
djangorestframework==3.16.1
djangorestframework-simplejwt==5.2.2
dnspython==2.7.0
email-validator==2.0.0
google-auth==2.40.3
google-genai==1.32.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
injector==0.22.0
Mako==1.3.10
MarkupSafe==3.0.2
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==2.10.1
python-dotenv==1.1.1
requests==2.32.5
rsa==4.9.1
ruff==0.12.11
setuptools==67.6.1
sniffio==1.3.1
soupsieve==2.8
sqlparse==0.5.3
tenacity==9.1.2
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
websockets==15.0.1
wikipedia==1.4.0
